[INTV] Fixed freeze when loading "stonix" game (it doesn't work though). Fixed V flag in CPU.
git-svn-id: svn://messdev.no-ip.org/mess@6345 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/machine/intv.c
@@ -328,7 +328,7 @@ static int intv_load_rom_file(const device_config *image)
 	UINT8 start_seg;
 	UINT8 end_seg;
 
-	UINT16 current_address;
+	UINT32 current_address;
 	UINT32 end_address;
 
 	UINT8 high_byte;
